create TeamInfo
